AI Agent/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
openai==1.63.2
